Data.Conduit.Tar.Types

Description

Module contains all the types necessary for tarball processing.

Documentation

data Header

Constructors

Header

Fields

headerOffset :: !FileOffset

headerPayloadOffset :: !FileOffset

headerFileNameSuffix :: !ShortByteString

headerFileMode :: !CMode

headerOwnerId :: !UserID

headerGroupId :: !GroupID

headerPayloadSize :: !FileOffset

headerTime :: !EpochTime

headerLinkIndicator :: !Word8

headerLinkName :: !ShortByteString

headerMagicVersion :: !ShortByteString

headerOwnerName :: !ShortByteString

headerGroupName :: !ShortByteString

headerDeviceMajor :: !DeviceID

headerDeviceMinor :: !DeviceID

headerFileNamePrefix :: !ShortByteString

data TarChunk

Constructors

ChunkHeader Header

ChunkPayload !FileOffset !ByteString

ChunkException TarException

data TarException

This the the exception type that is used in this module.

More constructors are susceptible to be added without bumping the major version of this module.

Constructors

NoMoreHeaders

UnexpectedPayload !FileOffset

IncompleteHeader !FileOffset

IncompletePayload !FileOffset !ByteCount

ShortTrailer !FileOffset

BadTrailer !FileOffset

InvalidHeader !FileOffset

BadChecksum !FileOffset

FileTypeError !FileOffset !Char !String

UnsupportedType !FileType

data TarCreateException

Constructors

FileNameTooLong !FileInfo

TarCreationError !String

data FileType

Constructors

FTNormal

FTHardLink !ByteString

FTSymbolicLink !ByteString

FTCharacterSpecial

FTBlockSpecial

FTDirectory

FTFifo

FTOther !Word8

data FileInfo

Constructors

FileInfo

Fields

filePath :: !ByteString

File path.

fileUserId :: !UserID

Unix user id.

fileUserName :: !ByteString

Unix user name.

fileGroupId :: !GroupID

Unix group id.

fileGroupName :: !ByteString

Unix group name.

fileMode :: !FileMode

Unix file permissions

fileSize :: !FileOffset

File size

fileType :: !FileType

File type. FTNormal, FTHardLink (@since 0.3.0), FTSymbolicLink and FTDirectory are the only ones supported for now

fileModTime :: !EpochTime

File modification timestamp

type FileOffset = COff

type ByteCount = CSize

type UserID = CUid

type GroupID = CGid

type DeviceID = CDev

type EpochTime = CTime

newtype CUid

Constructors

CUid Word32

newtype CGid

Constructors

CGid Word32

encodeFilePath :: FilePath -> ByteString

Convert FilePath into a UTF-8 encoded ByteString

decodeFilePath :: ByteString -> FilePath

Convert UTF-8 encoded ByteString back into the FilePath.

getFileInfoPath :: FileInfo -> FilePath

Get the FilePath.